Move tags widget out of WidgetDerivatives into own file
